+++ tests/test_integration_session.py
@@ -1,0 +1,7 @@
+import pytest
+xfail = pytest.mark.xfail
+
+
+@xfail()
+def test_run_cmd():
+    raise NotImplementedError()