+++ winrm/__init__.py
@@ -1,9 +1,8 @@
-from __future__ import unicode_literals
 import re
-from base64 import b64encode
-import xml.etree.ElementTree as ET
 
-from winrm.protocol import Protocol
+from winrm.psrp.psrp import PsrpClient
+from winrm.wsmv.wsmv import WsmvClient
+
 
 # Feature support attributes for multi-version clients.
 # These values can be easily checked for with hasattr(winrm, "FEATURE_X"),
@@ -12,106 +11,60 @@ from winrm.protocol import Protocol
 FEATURE_SUPPORTED_AUTHTYPES=['basic', 'certificate', 'ntlm', 'kerberos', 'plaintext', 'ssl', 'credssp']
 FEATURE_READ_TIMEOUT=True
 FEATURE_OPERATION_TIMEOUT=True
-
-class Response(object):
-    """Response from a remote command execution"""
-    def __init__(self, args):
-        self.std_out, self.std_err, self.status_code = args
-
-    def __repr__(self):
-        # TODO put tree dots at the end if out/err was truncated
-        return '<Response code {0}, out "{1}", err "{2}">'.format(
-            self.status_code, self.std_out[:20], self.std_err[:20])
-
+FEATURE_PSRP_CLIENT=True
+FEATURE_WSMV_CLIENT=True
 
 class Session(object):
-    # TODO implement context manager methods
-    def __init__(self, target, auth, **kwargs):
-        username, password = auth
-        self.url = self._build_url(target, kwargs.get('transport', 'plaintext'))
-        self.protocol = Protocol(self.url,
-                                 username=username, password=password, **kwargs)
 
-    def run_cmd(self, command, args=()):
-        # TODO optimize perf. Do not call open/close shell every time
-        shell_id = self.protocol.open_shell()
-        command_id = self.protocol.run_command(shell_id, command, args)
-        rs = Response(self.protocol.get_command_output(shell_id, command_id))
-        self.protocol.cleanup_command(shell_id, command_id)
-        self.protocol.close_shell(shell_id)
-        return rs
+    def __init__(self, **transport_opts):
+        self.transport_opts = transport_opts
+        self.transport_opts['endpoint'] = self._build_url(self.transport_opts['endpoint'])
+        if 'auth_method' not in self.transport_opts.keys():
+            self.transport_opts['auth_method'] = 'basic'
 
-    def run_ps(self, script):
-        """base64 encodes a Powershell script and executes the powershell
-        encoded script command
-        """
-        # must use utf16 little endian on windows
-        encoded_ps = b64encode(script.encode('utf_16_le')).decode('ascii')
-        rs = self.run_cmd('powershell -encodedcommand {0}'.format(encoded_ps))
-        if len(rs.std_err):
-            # if there was an error message, clean it it up and make it human
-            # readable
-            rs.std_err = self._clean_error_msg(rs.std_err)
-        return rs
-
-    def _clean_error_msg(self, msg):
-        """converts a Powershell CLIXML message to a more human readable string
-        """
-        # TODO prepare unit test, beautify code
-        # if the msg does not start with this, return it as is
-        if msg.startswith("#< CLIXML\r\n"):
-            # for proper xml, we need to remove the CLIXML part
-            # (the first line)
-            msg_xml = msg[11:]
-            try:
-                # remove the namespaces from the xml for easier processing
-                msg_xml = self._strip_namespace(msg_xml)
-                root = ET.fromstring(msg_xml)
-                # the S node is the error message, find all S nodes
-                nodes = root.findall("./S")
-                new_msg = ""
-                for s in nodes:
-                    # append error msg string to result, also
-                    # the hex chars represent CRLF so we replace with newline
-                    new_msg += s.text.replace("_x000D__x000A_", "\n")
-            except Exception as e:
-                # if any of the above fails, the msg was not true xml
-                # print a warning and return the orignal string
-                # TODO do not print, raise user defined error instead
-                print("Warning: there was a problem converting the Powershell"
-                      " error message: %s" % (e))
-            else:
-                # if new_msg was populated, that's our error message
-                # otherwise the original error message will be used
-                if len(new_msg):
-                    # remove leading and trailing whitespace while we are here
-                    msg = new_msg.strip()
-        return msg
-
-    def _strip_namespace(self, xml):
-        """strips any namespaces from an xml string"""
+    def run_cmd(self, command, arguments=()):
+        protocol = WsmvClient(self.transport_opts)
         try:
-            p = re.compile("xmlns=*[\"\"][^\"\"]*[\"\"]")
-            allmatches = p.finditer(xml)
-            for match in allmatches:
-                xml = xml.replace(match.group(), "")
-            return xml
-        except Exception as e:
-            raise Exception(e)
+            protocol.open_shell()
+            output = protocol.run_command(command, arguments)
+        finally:
+            protocol.close_shell()
+
+        return output
+
+    def run_ps(self, command, parameters=()):
+        protocol = PsrpClient(self.transport_opts)
+        try:
+            protocol.open_shell()
+            output = protocol.run_command(command, parameters=parameters)
+        finally:
+            protocol.close_shell()
+
+        return output
+
 
     @staticmethod
-    def _build_url(target, transport):
+    def _build_url(target):
         match = re.match(
-            '(?i)^((?P<scheme>http[s]?)://)?(?P<host>[0-9a-z-_.]+)(:(?P<port>\d+))?(?P<path>(/)?(wsman)?)?', target)  # NOQA
+            '(?i)^((?P<scheme>http[s]?)://)?(?P<host>[0-9a-z-_.]+)(:(?P<port>\d+))?(?P<path>(/)?(wsman)?)?', target)
         scheme = match.group('scheme')
-        if not scheme:
-            # TODO do we have anything other than HTTP/HTTPS
-            scheme = 'https' if transport == 'ssl' else 'http'
         host = match.group('host')
         port = match.group('port')
-        if not port:
-            port = 5986 if transport == 'ssl' else 5985
         path = match.group('path')
+
+        if not scheme:
+            if port:
+                if port == '5985':
+                    scheme = 'http'
+                else:
+                    scheme = 'https'
+            else:
+                scheme = 'https'
+        if not port:
+            if scheme == 'https':
+                port = 5986
+            else:
+                port = 5985
         if not path:
             path = 'wsman'
         return '{0}://{1}:{2}/{3}'.format(scheme, host, port, path.lstrip('/'))